requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
decorator==5.1.1
dnspython==2.2.1
email-validator==1.1.3
executing==0.8.3
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
